Add a type for errors
We still should discuss having a either a sum type or other affordance
for the case of parsing a response body which could be an error. You
can't just simply parse Either ESError a because aeson for some reason
provides an instance for (FromJSON a, FromJSON b) => FromJSON (Either a
b) which is highly Haskell-centric.