Expand test case for sub aggregation parsing
I'm not sure about the solution. We may have to take this out of a
typeclass and put it into a function that can keep track on if its in a
sub aggregation. One of the main issues is that the *user* must specify
buckets by name in their own code. ES throws in its own keys alongside
the bucket and there's no way bloodhound can distinguish the
two. Keeping a blacklist of known metadata keys will vary across
versions. Preemptively trying to parse the sub aggregation values could
be a performance issue and may hide other parsing bugs.

Refers to #86