Fix quickcheck failure due to ambiguous parse
Due to the way node attr filters are serialized (attr name as an object
key), its possible to express a value in haskell that will parse
ambiguously. If we wanted to move the invariant into the types we could
use a map I guess but whatever.

I ran this updated test through 100,000 tests and we're all green.